Добавлен тест testCodeFromURL
