+++ ImageFeedTests/WebViewViewControllerSpy.swift
@@ -1,0 +1,19 @@
+import Foundation
+import ImageFeed
+
+final class WebViewViewControllerSpy: WebViewViewControllerProtocol {
+    var loadRequest: Bool = false
+    var presenter: WebViewPresenterProtocol?
+    
+    func load(request: URLRequest) {
+        loadRequest = true
+    }
+    
+    func setProgressValue(_ newValue: Float) {
+        
+    }
+    
+    func setProgressHidden(_ isHidden: Bool) {
+        
+    }
+}
